### 2024-03-26 #6541 @kyleniemeyer
Hello @hellemo, thanks for your submission to JOSS. Unfortunately, we do not have an editor available in this area to handle this at the moment, so I have to put this in our waitlist until someone frees up.

In the meantime, any reviewer suggestions would be welcome.
