### 2024-03-26 #6541 @lrnv
@hellemo Greetings! I'll be your editor for this paper, nice to meet you!

Would you be able to recommend a few potential reviewers? You may e.g. take a look at Joss's reviewer list [there](https://reviewers.joss.theoj.org/), or propose anyone else that would be fitted (e.g. at least who know Julia and/or energy models).
