+++ src/utils.jl
@@ -101,32 +101,82 @@ function sort_types(types_list::Dict)
 end
 
 """
-    multiple(t_inv, t)
+    previous_level(
+        m,
+        n::Storage,
+        prev_pers::PrevPeriods,
+        t::TS.TimePeriod,
+        modeltype::EnergyModel,
+    )
 
-Provide a simplified function for returning the combination of the functions
-duration(t) * multiple_strat(t_inv, t) * probability(t)
+Returns the level used as previous level of a `Storage` node depending on the type of
+[`PrevPeriods`](@ref).
+
+The basic functionality is used in the case when the previous operational period is a
+`TimePeriod`, in which case it just returns the previous operational period.
 """
-multiple(t_inv, t) = duration(t) * multiple_strat(t_inv, t) * probability(t)
-
-# function previous_level(
-#     m,
-#     n::Storage,
-#     prev_pers::PrevPeriods{Nothing, Nothing, Nothing},
-#     t::TS.TimePeriod,
-#     modeltype::EnergyModel,
-# )
-#     @info "Frist operational period in the first strategic period"
-
-# end
-
 function previous_level(
     m,
     n::Storage,
-    prev_pers::PrevPeriods{<:nt, Nothing, Nothing},
+    prev_pers::PrevPeriods,
     t::TS.TimePeriod,
     modeltype::EnergyModel,
 )
 
+    # Previous storage level, as there are no changes
+    return @expression(m, m[:stor_level][n, prev_pers.op])
+end
+"""
+    previous_level(
+        m,
+        n::Storage,
+        prev_pers::PrevPeriods{<:nt, Nothing, Nothing, <:TS.TimePeriod},
+        t::TS.TimePeriod,
+        modeltype::EnergyModel,
+    )
+
+When the previous operational and representative period are `Nothing` and the last period
+is a `TimePeriod`, then the time structure *does not* include `RepresentativePeriods` and
+the cyclic default constraints returns a the last operational period of the strategic
+period.
+"""
+function previous_level(
+    m,
+    n::Storage,
+    prev_pers::PrevPeriods{<:nt, Nothing, Nothing, <:TS.TimePeriod},
+    t::TS.TimePeriod,
+    modeltype::EnergyModel,
+)
+    # Return the previous storage level based on cyclic constraints
+    return @expression(
+        m,
+        # Storage level in previous operational period
+        m[:stor_level][n, prev_pers.last]
+    )
+end
+"""
+    previous_level(
+        m,
+        n::Storage,
+        prev_pers::PrevPeriods{<:nt, Nothing, Nothing, <:TS.AbstractRepresentativePeriod},
+        t::TS.TimePeriod,
+        modeltype::EnergyModel,
+    )
+
+When the previous operational and representative period are `Nothing` and the last period
+is an `AbstractRepresentativePeriod`, then the time structure *does* include
+`RepresentativePeriods`.
+
+The cyclic default constraints returns the value at the end of the *last* representative
+period while accounting for the number of  repetitions of the representative period.
+"""
+function previous_level(
+    m,
+    n::Storage,
+    prev_pers::PrevPeriods{<:nt, Nothing, Nothing, <:TS.AbstractRepresentativePeriod},
+    t::TS.TimePeriod,
+    modeltype::EnergyModel,
+)
     # Return the previous storage level based on cyclic constraints when representative
     # periods are included
     return @expression(
@@ -137,57 +187,27 @@ function previous_level(
         # Increase in previous representative period
         m[:stor_level_Δ_rp][n, prev_pers.last]
     )
-
[…]
