+++ src/constraint_functions.jl
@@ -148,9 +148,6 @@ function constraints_level(m, n::Storage, 𝒯, 𝒫, modeltype::EnergyModel)
     constraints_level_aux(m, n, 𝒯, 𝒫, modeltype)
 
     # Mass/energy balance constraints for stored energy carrier.
-    # for t_inv ∈ 𝒯ᴵⁿᵛ
-    #     constraints_level_sp(m, n, t_inv, 𝒫, modeltype)
-
     for (t_inv_prev, t_inv) ∈ withprev(𝒯ᴵⁿᵛ)
         # Calculation of the operational period for the cyclic constraints
         last_per = last(collect(t_inv))
@@ -158,7 +155,7 @@ function constraints_level(m, n::Storage, 𝒯, 𝒫, modeltype::EnergyModel)
         # Creation of the iterator and call of the iterator function -
         # The representative period is initiated with the current investment period to allows
         # for dispatching on it.
-        prev_pers = PrevPeriods(t_inv_prev, t_inv,  nothing, last_per);
+        prev_pers = PrevPeriods(t_inv_prev, nothing,  nothing, last_per);
         ts = t_inv.operational
         constraints_level_iterate(m, n, prev_pers, t_inv, ts, modeltype)
     end
@@ -166,12 +163,12 @@ end
 
 
 """
-    constraints_level_aux(m, n::RefStorage{S}, 𝒯, 𝒫, modeltype::EnergyModel) where {S<:ResourceCarrier}
+    constraints_level_aux(m, n::Storage, 𝒯, 𝒫, modeltype::EnergyModel)
 
 Function for creating the Δ constraint for the level of a reference storage node with a
 `ResourceCarrier` resource.
 """
-function constraints_level_aux(m, n::RefStorage{S}, 𝒯, 𝒫, modeltype::EnergyModel) where {S<:ResourceCarrier}
+function constraints_level_aux(m, n::Storage, 𝒯, 𝒫, modeltype::EnergyModel)
     # Declaration of the required subsets
     p_stor = storage_resource(n)
 
@@ -211,6 +208,23 @@ function constraints_level_aux(m, n::RefStorage{S}, 𝒯, 𝒫, modeltype::Energ
     @constraint(m, [t ∈ 𝒯, p_em ∈ 𝒫ᵉᵐ], m[:emissions_node][n, t, p_em] == 0)
 end
 
+"""
+    constraints_level_iterate(
+        m,
+        n::Storage,
+        prev_pers::PrevPeriods,
+        per,
+        ts::RepresentativePeriods,
+        modeltype::EnergyModel,
+    )
+
+Iterate through the individual time structures of a `Storage` node. This iteration function
+should in general allow for all necessary functionality for incorporating modifications.
+
+In the case of `RepresentativePeriods`, this is achieved through calling the function
+[`constraints_level_rp`](@ref) to introduce, _e.g._, cyclic constraints as it is in the
+default case.
+ """
 function constraints_level_iterate(
     m,
     n::Storage,
@@ -223,11 +237,14 @@ function constraints_level_iterate(
     𝒯ʳᵖ = repr_periods(per)
     last_per = last(𝒯ʳᵖ)
 
-    # Constraint for the cycle storage constraints when including representative periods
-    constraints_level_cyclic_rp(m, n, per, ts, modeltype)
+    # Constraint for additional, node specific constraints for representative periods
+    constraints_level_rp(m, n, per, ts, modeltype)
 
-    # Constraint that the total change has to be 0
-    @constraint(m, sum(m[:stor_level_Δ_rp][n, t_rp] for t_rp ∈ 𝒯ʳᵖ) == 0)
[…]
