### 2024-04-09 e449aa2
Modification in previous_level for reduction in code
