+++ src/structures/misc.jl
@@ -2,9 +2,9 @@
 
 
 nt = Union{Nothing, TS.TimePeriod, TS.TimeStructure{T}} where {T}
-struct PrevPeriods{S<:nt, T<:nt, U<:nt,  V<:nt}
+struct PrevPeriods{S<:nt, T<:nt, U<:nt}
     sp::S
     rp::T
     op::U
-    last::V
+    last::nt
 end
