### 2024-04-14 #6619 @editorialbot
[REVIEW]: EnergyModelsX: Flexible Energy Systems Modelling with Multiple Dispatch
**Submitting author:** @hellemo (<a href="http://orcid.org/0000-0001-5958-9794">Lars Hellemo</a>)
**Repository:** https://github.com/EnergyModelsX/EnergyModelsBase.jl
**Branch with paper.md** (empty if default branch): paper_submission/joss
**Version:** v0.6.8
