### 2024-04-14 EnergyModelsX/EnergyModelsBase.jl#21 @odow
JOSS Review
x-ref https://github.com/openjournals/joss-reviews/issues/6619

## Checklist items

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

You could add a paragraph to the documentation, or a CONTRIBUTING.md file.

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)?

I would add:
 * https://github.com/GenXProject/GenX.jl
 * https://github.com/TulipaEnergy/TulipaEnergyModel.jl

## Other

 - [x] I think the paper would benefit from a pretty picture or an example. I don't have any concrete suggestion, but could you make a figure with a plot of input time series and an output of various dispatch/cost/flows, etc?
 - [x] https://github.com/EnergyModelsX/EnergyModelsGeography.jl/pull/15
