### 2024-04-16 c5f7d16
Added potential for negative emission technologies (#22)
