+++ src/structures/misc.jl
@@ -6,5 +6,9 @@ struct PrevPeriods{S<:nt, T<:nt, U<:nt}
     sp::S
     rp::T
     op::U
-    last::nt
+end
+
+struct CyclicPeriods{S<:nt}
+    last::S
+    current::S
 end
