### 2024-04-16 6aaeef5
Added a contribution section to the documentation
