### 2024-04-16 EnergyModelsX/EnergyModelsBase.jl#21 @JulStraus
First of all, thanks for agreeing to review the paper. We are currently working on incorporating the individual comments:

1.  The individual links in the different packages were updated to not point towards internal gitlab pages.
2. We added a contribution section to the documentation of `EnergyModelsBase`. Once this is accepted, we will update the other packages correspondingly.
3. We will look into the references and update the list accordingly.

@hellemo and I are currently discussing a potential figure, as we agree that it can improve the paper. We are however uncertain, whether a results figure is what we aim for in this paper. We will however update the paper soon with a figure.
