### 2024-04-16 EnergyModelsX/EnergyModelsBase.jl#21 @odow
> whether a results figure is what we aim for in this paper

Yeah I understand. I guess something more indicative of what this is/could be used for. Currently, the paper is fairly abstract in the sense "you can use it to build things" but "you can use it to build a model of hydrogen generation and transshipments in Scandinavia with wind uncertainty" is much more concrete. Dunno. Just a pretty picture.
