### 2024-04-16 9cc4c4d
Add references to GenX and Tulipa, fix SpineOpt DOI
