+++ docs/joss/paper.md
@@ -53,7 +53,7 @@ See @bodal2024hydrogen for an example application of `EnergyModelsX`.
 The increasing share of renewable energy generation and importance of sector coupling increases the complexity of energy systems, and makes the modelling of these systems more challenging.
 To meet the demand of energy modelers, energy system models need ever increasing flexibility to analyse the energy systems of tomorrow [@fodstad2022next].
 While large scale models like TIMES [@times] and GENeSYS-MOD [@genesysmod] are important for modelling large energy systems, they lack the potential for simple modifications in technology descriptions as well as simple incorporation of region specific constraints.
-[SpineOpt](https://github.com/spine-tools/SpineOpt.jl/tree/master) [@SpineOpt] offers the user with the flexibility, but the monolithic approach of including all functionality in a single package reduces the understandability of the code.
+[SpineOpt](https://github.com/spine-tools/SpineOpt.jl/tree/master) [@SpineOpt] offers the user with the flexibility, but the monolithic approach of including all functionality in a single package reduces the understandability of the code. [GenX](https://github.com/GenXProject/GenX.jl) [@GenX] and [Tulipa Energy Model](https://github.com/TulipaEnergy/TulipaEnergyModel.jl) [@Tulipa] are other recent energy system models developed in Julia with similar goals to EnergyModelsX, but with less focus on extensibility and alternative technology formulations.  
 
 [EnergyModelsX](https://github.com/EnergyModelsX/) is a modular energy-system modelling framework designed to give modelers a high level of flexibility.
 The time resolution is decoupled from the technology descriptions by the application of [TimeStruct](https://github.com/sintefore/TimeStruct.jl) [@TimeStruct], facilitating the support of a wide range of time structures with different temporal resolution and to support operational uncertainty.
