+++ docs/joss/paper.bib
@@ -55,7 +55,7 @@ volume = {43},
 pages = {100902},
 year = {2022},
 issn = {2211-467X},
-doi = {https://doi.org/10.1016/j.esr.2022.100902},
+doi = {10.1016/j.esr.2022.100902},
 url = {https://www.sciencedirect.com/science/article/pii/S2211467X22000955},
 author = {Maren Ihlemann and Iasonas Kouveliotis-Lysikatos and Jiangyi Huang and Joseph Dillon and Ciara O’Dwyer and Topi Rasku and Manuel Marin and Kris Poncelet and Juha Kiviluoma},
 keywords = {Open source tool, Energy system modelling, Energy system analysis, Integrated energy systems, Investment planning, Sector coupling},
@@ -79,4 +79,24 @@ ARTICLE-NUMBER = {1468},
 URL = {https://www.mdpi.com/1996-1073/10/10/1468},
 ISSN = {1996-1073},
 DOI = {10.3390/en10101468}
+}
+
+@misc{GenX,
+  author       = {Jenkins, Jesse and Sepulveda, Nestor and Mallapragada, Dharik and Patankar, Neha and Schwartz, Aaron and Schwartz, Jacob and Chakrabarti, Sambuddha and Xu, Qingyu and Morris, Jack and Sepulveda, Nestor},
+  title        = {GenX},
+  month        = Apr,
+  year         = 2024,
+  doi          = {https://zenodo.org/doi/10.5281/zenodo.10846069},
+  publisher    = {Zenodo},
+  url          = {https://zenodo.org/records/10910030}
+}
+
+@misc{Tulipa,
+  author       = {Tejada-Arango, Diego and Morales-España, Germán and Clisby, Lauren and Wang, Ni and Soares Siqueira, Abel and Ali, Suvayu and Soucasse, Laurent and Neustroev, Grigory},
+  title        = {Tulipa Energy Model},
+  month        = Mar,
+  year         = 2024,
+  doi          = {10.5281/zenodo.8363262},
+  publisher    = {Zenodo},
+  url          = {https://zenodo.org/records/10895406}
 }
