### 2024-04-17 8677ef6
Modified storage for better cyclic representation
The Storage node was changed to account for different cyclic behaviours.
This allows potential users to avoid reimplemeting the level balance for
Storage nodes
