+++ src/structures/node.jl
@@ -2,6 +2,45 @@
 abstract type Node end
 Base.show(io::IO, n::Node) = print(io, "n_$(n.id)")
 
+""" `StorageBehavior` as supertype for individual storage behaviours."""
+abstract type StorageBehavior end
+
+""" `Accumulating` as supertype for an accumulating storage level."""
+abstract type Accumulating <: StorageBehavior end
+
+""" `Cyclic` as supertype for a cyclic storage level."""
+abstract type Cyclic <: StorageBehavior end
+
+"""
+    AccumulatingEmissions <: Accumulating
+
+`StorageBehavior` which accumulates all inflow witin a strategic period.
+`AccumulatingEmissions` allows as well to serve as a `ResourceEmit` emission point to
+represent a soft constraint on storing the captured emissions.
+"""
+struct AccumulatingEmissions <: Accumulating end
+
+"""
+    CyclicStrategic <: Cyclic
+
+`StorageBehavior` in which the the cyclic behaviour is achieved within a strategic period.
+This implies that the initial level in individual representative periods can be different
+when using `RepresentativePeriods`
+"""
+struct CyclicStrategic <: Cyclic end
+
+"""
+    CyclicRepresentative <: Cyclic
+
+`StorageBehavior` in which cyclic behaviour is achieved within the lowest time structure
+excluding operational times.
+
+In the case of `TwoLevel{SimpleTimes}`, this approach is similar to `CyclicStrategic`.\n
+In the case of `TwoLevel{RepresentativePeriods{SimpleTimes}}`, this approach differs from
+`CyclicStrategic` as the cyclic constraint is enforeced within each representative period.
+"""
+struct CyclicRepresentative <: Cyclic end
+
 """ `Source` node with only output."""
 abstract type Source <: Node end
 """ `NetworkNode` node with both input and output."""
@@ -9,7 +48,7 @@ abstract type NetworkNode <: Node end
 """ `Sink` node with only input."""
 abstract type Sink <: Node end
 """ `Storage` node with level."""
-abstract type Storage <: NetworkNode end
+abstract type Storage{T<:StorageBehavior} <: NetworkNode end
 """ `Availability` node as routing node."""
 abstract type Availability <: NetworkNode end
 
@@ -96,7 +135,7 @@ GenAvailability(id, 𝒫::Vector{<:Resource}) = GenAvailability(id, 𝒫, 𝒫)
 """ A reference `Storage` node.
 
 This node is designed to store either a `ResourceCarrier` or a `ResourceEmit`.
-It is designed as a composite type to automatically distinguish between these two.
+It is designed as a composite type to differentiate between different cyclic behaviour.
 
 # Fields
 - **`id`** is the name/identifier of the node.\n
@@ -104,95 +143,35 @@ It is designed as a composite type to automatically distinguish between these tw
 - **`stor_cap::TimeProfile`** is the installed storage capacity, that is e.g. energy or mass.\n
 - **`opex_var::TimeProfile`** is the variational operational costs per energy unit stored.\n
 - **`opex_fixed::TimeProfile`** is the fixed operational costs.\n
-- **`stor_res::T`** is the stored `Resource`.\n
+- **`stor_res::Resource`** is the stored `Resource`.\n
 - **`input::Dict{<:Resource, <:Real}`** are the input `Resource`s with conversion value `Real`.
 - **`output::Dict{<:Resource, <:Real}`** are the generated `Resource`s with conversion value `Real`. \
 Only relevant for linking and the stored `Resource`.\n
[…]
