### 2024-04-18 bf1f1b7
Added a contribution section to the documentation (#23)
