+++ docs/src/how-to/contribute.md
@@ -1,0 +1,116 @@
+# Contributing to EnergyModelsBase
+
+Contributing to `EnergyModelsBase` can be achieved in several different ways.
+
+## Creating new extensions
+
+The main focus of `EnergyModelsBase` is to provide an easily extendable energy system optimization modelling framework.
+Hence, a first approach to contributing to `EnergyModelsBase` is to create a new package with, _e.g._, the introduction of new node descriptions.
+
+!!! tip
+    If you are uncertain how you could incorporate new nodal descriptions, take a look at [`EnergyModelsRenewableProducers`](https://github.com/EnergyModelsX/EnergyModelsRenewableProducers.jl).
+    The package is maintained by the developers of `EnergyModelsBase`.
+    Hence, it provides you with ideas on how we think it is best to develop new node descriptions.
+
+## File a bug report
+
+Another approach to contributing to `EnergyModelsBase` is through filing a bug report as an _[issue](https://github.com/EnergyModelsX/EnergyModelsBase.jl/issues/new)_ when unexpected behaviour is occuring.
+
+When filing a bug report, please follow the following guidelines:
+
+1. Be certain that the bug is a bug and originating in `EnergyModelsbase`:
+    - If the problem is within the results of the optimization problem, please check first that the nodes are correctly linked with each other.
+      Frequently, missing links (or wrongly defined links) restrict the transport of energy/mass.
+      If you are certain that all links are set correctly, it is most likely a bug in `EnergyModelsBase` and should be reported.
+    - If the problem occurs in model construction, it is most likely a bug in `EnergyModelsBase` and should be reported.
+    - If the problem is only appearing for specific solvers, it is most likely not a bug in `EnergyModelsBase`, but instead a problem of the solver wrapper for `MathOptInterface`. In this case, please contact the developers of the corresponding solver wrapper.
+2. Label the issue as bug,
+3. Provide a minimum working example of a case in which the bug occurs.
+
+!!! note
+    We are aware that certain design choices within `EnergyModelsBase` can lead to method ambiguities.
+    Our aim is to extend the documentation to improve the description on how to best extend the base functionality as well as which caveats can occur.
+
+    In order to improve the code, we welcome any reports of potential method ambiguities to help us improving the structure of the framework.
+
+## Feature requests
+
+Although `EnergyModelsBase` was designed with the aim of flexibility, it sometimes still requires additional features to account for potential extensions.
+Feature requests can be achieved through two approaches:
+
+1. Create an issue describing the aim of the feature request and
+2. Incorporate the feature request through a fork of the repository.
+
+!!! note
+    `EnergyModelsBase` should not include everything.
+
+    The aim of the framework is to be lightweight and extensible by the user.
+    Hence, feature requests should only include basic requirements for the core structure, and not, _e.g._, the description of new technologies.
+    These should be developed outside of `EnergyModelsBase`.
+
+### [Create an Issue](@id create_issue)
+
+Creating a new _[issue](https://github.com/EnergyModelsX/EnergyModelsBase.jl/issues/new)_ for a feature request is our standard approach for extending `EnergyModelsBase`.
+Due to the extendibility of `EnergyModelsBase`, it is not necessarily straight forward to understand how to best incorporate required features into the framework without breaking other packages.
+
+When creating a new issue as feature request, please follow the the following guidelines.
+
+1. **Reason for the feature**: Please describe the reasoning for the feature request. What functionality do you require in the core structure of `EnergyModelsBase`?
+2. **Required outcome**: What should be the outcome when including the feature and what should be the minimum requirements of the outcome?
+3. **Potential solutions**: Describe alternatives you consider. This step is not necessarily required, but can be helpful for identifying potential solutions.
+
+### Incorporating the feature requests through a fork
+
[…]
