+++ NEWS.md
@@ -1,10 +1,11 @@
 # Release notes
 
-## Unversioned
+## Version 0.6.8 (2024-04-18)
 
 * Added potential for negative emissions.
   This change requires the user to always constrain the variable `emissions_node`, if it is defined by the user.
-  By default, this is achieved in the developed packages through `EmissionsData` or additional bounds.
+  By default, this is achieved in the developed packages through `EmissionsData` or the addition of additional bounds on
+  the variable `:emissions_node` through the JuMP function [`set_lower_bound`](https://jump.dev/JuMP.jl/stable/api/JuMP/#set_lower_bound).
 * Provided a contribution section in the documentation.
 
 ## Version 0.6.7 (2024-03-21)
