+++ docs/src/index.md
@@ -36,6 +36,17 @@ Pages = [
     "manual/constraint-functions.md",
     "manual/data-functions.md",
     "manual/simple-example.md",
+    "manual/simple-example.md",
+]
+```
+
+## How to guides
+
+```@contents
+Pages = [
+    "how-to/create-new-node.md",
+    "how-to/utilize-timestruct.md",
+    "how-to/contribute.md",
 ]
 ```
 
