### 2024-04-18 09f9947
Docs/fix typos (#24)
* Fixed minor errors in contribution section

* Added how to pages to outline
