### 2024-04-18 a324adf
Removed \texttt in docstrings (#25)
