+++ v0.6.8/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"how-to/utilize-timestruct/#utilize_timestruct","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"","category":"section"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"EnergyModelsBase uses for the description of time the package TimeStruct. TimeStruct offers a large variety of different options that can appear to be overwhelming, when first exposed to them. Hence, it is important to highlight how it works and which parameters you would want to analyse.","category":"page"},{"location":"how-to/utilize-timestruct/#Structures-for-time-description","page":"Utilize TimeStruct","title":"Structures for time description","text":"","category":"section"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"TimeStruct introduces individual structures that are used for describing time. In the following introduction, the most important structures are explained. There are other structures, but these will be added once EnergyModelsbase supports their formulation.","category":"page"},{"location":"how-to/utilize-timestruct/#Operational-periods","page":"Utilize TimeStruct","title":"Operational periods","text":"","category":"section"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"Operational periods correspond to periods in which no investments are allowed. In general, you can imagine operational periods to be the individual hours you want to model, although it is not limited to hours. In each operational period, we have an optimal dispatch problem given the constraints. Operational periods are normally implemented using the type SimpleTimes which has the following structure:","category":"page"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"op_duration = 1  # Each operational period has a duration of 1\nop_number = 24   # There are in total 24 operational periods\noperational_periods = SimpleTimes(op_number, op_duration)","category":"page"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"In above's example, we assume that we have in total 24 operational periods which have the same duration each, i.e., a duration of 1. The unit in itself is not important, but it can be either if you consider that a duration of 1 corresponds to one hour. In this case, the operational periods would correspond to a full day.","category":"page"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"note: Note\nAll operational periods are continuous. This implies that one is not allowed to have jumps in the representative periods. This affects all \"dynamic\" constraints, that is, constraints where the current operational period is dependent on the previous operational period. In EnergyModesBase, this is only the case for the level balance in RefStorage. Representative periods allow for jumps between operational periods, as outlined below.","category":"page"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"Note that TimeStruct does not require that each operational period has the same length. Consider the following example:","category":"page"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","title":"Utilize TimeStruct","text":"op_duration = [4, 2, 1, 1, 2, 4, 2, 1, 1, 2, 4]\nop_number = length(op_duration)\noperational_periods = SimpleTimes(op_number, op_duration)\n\n# output\nSimpleTimes{Int64}(11, [4, 2, 1, 1, 2, 4, 2, 1, 1, 2, 4])","category":"page"},{"location":"how-to/utilize-timestruct/","page":"Utilize TimeStruct","
[…]
