### 2024-04-19 72393f6
Add figure with example of application Fix URL for SINTEF in Summary
