### 2024-04-19 EnergyModelsX/EnergyModelsBase.jl#21 @hellemo
Thanks for reviewing our submission, @odow !

As mentioned above, @JulStraus has updated the documentation with a new sections addressing contributions.

We have also updated the paper manuscript now with the following changes:
1. Added references to GenX and Tulipa
2. Added a figure with an example application (hydrogen infrastructure development in the North Sea)
3. Fixed the flagged DOI
4. Fixed the URL for link to SINTEF
