### 2024-04-22 cc139f7
Updated the documentation based for storage
The changes incorporated for Storage nodes did not include an update to
the documentation.
