### 2024-04-23 e7cb998
Added tests for CyclicReprepresentative behaviour
