+++ test/test_nodes.jl
@@ -224,7 +224,7 @@
     end
 end
 
-@testset "Test RefNetworkNode with various emission variables" begin
+@testset "Test RefNetworkNode with various EmissionData" begin
     # Resources used in the analysis
     NG = ResourceEmit("NG", 0.2)
     Power = ResourceCarrier("Power", 0.0)
@@ -371,7 +371,7 @@ end
         # Check that the total and strategic emissions are correctly calculated
         # - constraints_data(m, n::Node, 𝒯, 𝒫, modeltype, data::EmissionsEnergy)
         @test sum(value.(m[:emissions_node][net, t, CO2]) ≈
-                sum(value.(m[:flow_in][net, t, p]) * EMB.co2_int(p) for p ∈ inputs(net))
+                sum(value.(m[:flow_in][net, t, p]) * co2_int(p) for p ∈ inputs(net))
                 for t ∈ 𝒯, atol=TEST_ATOL) ≈
                     length(𝒯)
         @test sum(value.(m[:emissions_node][net, t, NG]) ≈ 0
@@ -406,7 +406,7 @@ end
         # Check that the total and strategic emissions are correctly calculated
         # - constraints_data(m, n::Node, 𝒯, 𝒫, modeltype, data::EmissionsProcess)
         @test sum(value.(m[:emissions_node][net, t, CO2]) ≈
-                sum(value.(m[:flow_in][net, t, p]) * EMB.co2_int(p) for p ∈ inputs(net)) +
+                sum(value.(m[:flow_in][net, t, p]) * co2_int(p) for p ∈ inputs(net)) +
                 value.(m[:cap_use][net, t]) * process_emissions(em_data, CO2, t)
                 for t ∈ 𝒯, atol=TEST_ATOL) ≈
                     length(𝒯)
@@ -443,7 +443,7 @@ end
         # Check that the total and strategic emissions are correctly calculated
         # - constraints_data(m, n::Node, 𝒯, 𝒫, modeltype, data::EmissionsProcess)
         @test sum(value.(m[:emissions_node][net, t, CO2]) ≈
-                sum(value.(m[:flow_in][net, t, p]) * EMB.co2_int(p) for p ∈ inputs(net)) +
+                sum(value.(m[:flow_in][net, t, p]) * co2_int(p) for p ∈ inputs(net)) +
                 value.(m[:cap_use][net, t]) * process_emissions(em_data, CO2, t)
                 for t ∈ 𝒯, atol=TEST_ATOL) ≈
                     length(𝒯)
@@ -480,8 +480,8 @@ end
         # Check that the total and strategic emissions are correctly calculated
         # - constraints_data(m, n::Node, 𝒯, 𝒫, modeltype, data::CaptureEnergyEmissions)
         @test sum(value.(m[:emissions_node][net, t, CO2]) ≈
-                sum(value.(m[:flow_in][net, t, p]) * EMB.co2_int(p) for p ∈ inputs(net)) *
-                (1 - EMB.co2_capture(em_data)) +
+                sum(value.(m[:flow_in][net, t, p]) * co2_int(p) for p ∈ inputs(net)) *
+                (1 - co2_capture(em_data)) +
                 value.(m[:cap_use][net, t]) * process_emissions(em_data, CO2, t)
                 for t ∈ 𝒯, atol=TEST_ATOL) ≈
                     length(𝒯) atol=TEST_ATOL
@@ -493,8 +493,8 @@ end
         # Test that the CO2 capture is calculated correctly
         # - constraints_data(m, n::Node, 𝒯, 𝒫, modeltype, data::CaptureEnergyEmissions)
         @test sum(value.(m[:flow_out][net, t, CO2]) ≈
[…]
