### 2024-04-23 178d43b
Added part support for OperationalScenarios
OperationalScenarios can be used with CyclicRepresentative storage
behaviour in which the last level in each scenario has to be the same
