### 2024-04-24 EnergyModelsX/EnergyModelsBase.jl#21 @odow
Just one other comment, you might want to add another citation to Bodal in the "Example application" section. It's currently just in the summary.
