### 2024-04-25 6bcbdad
Added reference to related @bodal2024hydrogen in Example section
