### 2024-04-25 EnergyModelsX/EnergyModelsBase.jl#21 @hellemo
Just updated with the following addition:

"See also @bodal2024hydrogen for a similar example."

The reason we didn't add it first was because the figure is from a separate study (in preparation), but they share many  properties in terms of modelling.
