+++ docs/joss/paper.bib
@@ -86,7 +86,7 @@ DOI = {10.3390/en10101468}
   title        = {GenX},
   month        = Apr,
   year         = 2024,
-  doi          = {https://zenodo.org/doi/10.5281/zenodo.10846069},
+  doi          = {doi/10.5281/zenodo.10846069},
   publisher    = {Zenodo},
   url          = {https://zenodo.org/records/10910030}
 }
