### 2024-05-06 #6619 @hellemo
v.0.6.8
https://zenodo.org/doi/10.5281/zenodo.11118908

@lrnv I registered `EnergyModelsBase` at Zenodo (see above), while the JOSS manuscript covers multiple repos. EnergyModelsBase is the same repo as is used for the JOSS manuscript. Let me know if this is ok.
