### 2024-05-07 542dcc9
Export changes - added export for functions for accessing fields in CyclicPeriods and   PreviousPeriods
