### 2024-05-08 #6619 @lrnv
@hellemo So apparently you CAN do a zenodo archive for several repo at once, but you have to do it manually (the zenodo/github integration does not do it for you).
