### 2024-05-09 #6619 @hellemo
@lrnv I just updated the comment above with the new Zenodo DOI: https://zenodo.org/doi/10.5281/zenodo.11164517 which includes a copy of the latest release of all the packages.
