### 2024-05-10 2e3395d
Fix OperationaProfiles for some tests using Matrix input where TimeStruct requires Vector.
