+++ examples/sink_source.jl
@@ -56,7 +56,7 @@ function generate_example_data()
         ),
         RefSink(
             "electricity demand",       # Node id
-            OperationalProfile([20 30 40 30]), # Demand in MW
+            OperationalProfile([20, 30, 40, 30]), # Demand in MW
             Dict(:surplus => FixedProfile(0), :deficit => FixedProfile(1e6)),
             # Line above: Surplus and deficit penalty for the node in EUR/MWh
             Dict(Power => 1),           # Energy demand and corresponding ratio
