### 2024-05-12 05a4dde
Implemented potential for discharge and charge * Contrary to previous implementation with rate in Storage, it is now   possible to have both charge and discharge capacities * In addition, the change allows for opex for both as well
