+++ src/legacy_constructor.jl
@@ -7,7 +7,7 @@ storage.
 See the documentation for further information. In this case, the key difference is that we
 changed the parameteric descriptions from the stored `Resource` to the behaviour of the
 `Storage` node as well as allowing for variable and fixed OPEX for both the level and the
-charge rate as well as allowing for the introduction of a discharge rate.
+charge rate as well as introducing a discharge rate capacity.
 """
 function RefStorage(
     id,
@@ -21,15 +21,18 @@ function RefStorage(
     data::Vector,
 )
 
-    @warn("The used implementation of a `RefStorage` will be discontinued in the near future. \
-    See the documentation for the new implementation using a parametric type describing \
-    the storage behaviour and the changes incorporated for level and charge through using \
-    the type `AbstractStorageParameters`.\n\
-    In practice, two changes have to be incorporated: \n 1. `RefStorage{AccumulatingEmissions}()` \
-    instead of `RefStorage` and \n 2. the application of `StorCapOpex(rate_cap, opex_var, opex_fixed)` \
-    as 2ⁿᵈ field as well as `StorCap(stor_cap)` as 3ʳᵈ field instead of using \
-    `rate_cap`, `stor_cap`, `opex_var`, and `opex_fixed` as 2ⁿᵈ-5ᵗʰ fields.\n\
-    It is recommended to update the existing implementation to the new version.")
+    @warn(
+        "The used implementation of a `RefStorage` will be discontinued in the near future. " *
+        "See the documentation for the new implementation using a parametric type describing " *
+        "the storage behaviour and the changes incorporated for level and charge through using " *
+        "the types `StorageBehavior` and `AbstractStorageParameters` in the scetion on " *
+        "_How to update your model to the latest versions_.\n" *
+        "In practice, two changes have to be incorporated: \n 1. `RefStorage{AccumulatingEmissions}()` " *
+        "instead of `RefStorage` and \n 2. the application of `StorCapOpex(rate_cap, opex_var, opex_fixed)` " *
+        "as 2ⁿᵈ field as well as `StorCap(stor_cap)` as 3ʳᵈ field instead of using " *
+        "`rate_cap`, `stor_cap`, `opex_var`, and `opex_fixed` as 2ⁿᵈ-5ᵗʰ fields.\n" *
+        "It is recommended to update the existing implementation to the new version."
+    )
 
     tmp = RefStorage{AccumulatingEmissions}(
         id,
@@ -53,15 +56,18 @@ function RefStorage(
     output::Dict{<:Resource,<:Real},
 )
 
-    @warn("The used implementation of a `RefStorage` will be discontinued in the near future. \
-    See the documentation for the new implementation using a parametric type describing \
-    the storage behaviour and the changes incorporated for level and charge through using \
-    the type `AbstractStorageParameters`.\n\
-    In practice, two changes have to be incorporated: \n1. `RefStorage{AccumulatingEmissions}()` \
-    instead of `RefStorage` and \n2. the application of `StorCapOpex(rate_cap, opex_var, opex_fixed)` \
-    as 2ⁿᵈ field as well as `StorCap(stor_cap)` as 3ʳᵈ field instead of using \
-    `rate_cap`, `stor_cap`, `opex_var`, and `opex_fixed` as 2ⁿᵈ-5ᵗʰ fields.\n\
-    It is recommended to update the existing implementation to the new version.")
+    @warn(
+        "The used implementation of a `RefStorage` will be discontinued in the near future. " *
+        "See the documentation for the new implementation using a parametric type describing " *
+        "the storage behaviour and the changes incorporated for level and charge through using " *
[…]
