### 2024-05-12 a226d44
Fixed errors in discharge and docs
