### 2024-05-13 EnergyModelsX/EnergyModelsBase.jl#27 @lrnv
[joss paper] remarks
Hi ! A few notes on the paper while i read it :

- The example section is abit weird, as, for me (I am not an expert in your field), i did not understand at all what your software did optimize based on what data to obtain the proposed graph. Could you enlarge a bit the section to detail clearly what are 1) Inputs, 2) aims, goals, algorithms, etc.. 3) Outputs ?
- While the zenodo archive and the repositories are all on MIT licence, the paper shows out CC By 4.0. This is not necessarily an issue but you might want to sync up (or not).

xref https://github.com/openjournals/joss-reviews/issues/6619
