### 2024-05-13 EnergyModelsX/EnergyModelsBase.jl#27 @hellemo
@lrnv Thanks for you comments!

We have updated the example section now, hope it is easier to understand now. We still try keeping it quite brief, as we were a bit concerned about the length (there tend to be a large number of details for all such case studies).

We discussed the license issue internally, and as we understand it the paper will be licensed under CC By 4.0, while the code is licensed under MIT, and that sounds reasonable to us, but feel free to elaborate if this is something that should be aligned.
