### 2024-05-13 EnergyModelsX/EnergyModelsBase.jl#27 @lrnv
The example is much easier to read now for me, thanks.

No reason to align the licenses, I just wanted you to consider the issue.

You are good to go now :)
