### 2024-05-13 #6619 @hellemo
If it's the missing DOI that is causing issues, we could consider switching the reference for TIMES to
https://doi.org/10.1007/s10287-007-0046-z but that is a quite a bit older  reference, and the manual (without DOI) is an often used reference for the TIMES model.
