### 2024-05-16 a033560
Fixed variables instead of constraining them
