### 2024-05-19 e7fb750
Different calculation of var OPEX for Storage - Reduced number of functions - Allowed to simplify the overall function call system
